% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{abort_sql}
\alias{abort_sql}
\title{AbortSQL}
\usage{
abort_sql(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\value{
integer
}
\description{
AbortSQL
}
