% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{add_band}
\alias{add_band}
\title{AddBand}
\usage{
add_band(x, datatype, options = NULL)
}
\arguments{
\item{x}{A GDALDataset object}

\item{datatype}{ANY}

\item{options}{character (default: NULL)}
}
\value{
integer
}
\description{
AddBand
}
