% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{as_mdarray}
\alias{as_mdarray}
\title{AsMDArray}
\usage{
as_mdarray(x, options = NULL)
}
\arguments{
\item{x}{A GDALDataset object}

\item{options}{character (default: NULL)}
}
\value{
GDALMDArray
}
\description{
AsMDArray
}
