% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{clear_statistics}
\alias{clear_statistics}
\title{ClearStatistics}
\usage{
clear_statistics(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\description{
ClearStatistics
}
