% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{commit_transaction}
\alias{commit_transaction}
\title{CommitTransaction}
\usage{
commit_transaction(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\value{
integer
}
\description{
CommitTransaction
}
