% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{create_mask_band}
\alias{create_mask_band}
\title{CreateMaskBand}
\usage{
create_mask_band(x, nFlags)
}
\arguments{
\item{x}{A GDALDataset object}

\item{nFlags}{integer}
}
\value{
integer
}
\description{
CreateMaskBand
}
