% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{delete_field_domain}
\alias{delete_field_domain}
\title{DeleteFieldDomain}
\usage{
delete_field_domain(x, name)
}
\arguments{
\item{x}{A GDALDataset object}

\item{name}{character}
}
\value{
logical
}
\description{
DeleteFieldDomain
}
