% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{delete_relationship}
\alias{delete_relationship}
\title{DeleteRelationship}
\usage{
delete_relationship(x, name)
}
\arguments{
\item{x}{A GDALDataset object}

\item{name}{character}
}
\value{
logical
}
\description{
DeleteRelationship
}
