% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{feature_count}
\alias{feature_count}
\title{Count the features of a layer}
\usage{
feature_count(x, force = TRUE)
}
\arguments{
\item{x}{A GDALLayer object.}

\item{force}{Count them by reading the layer when the driver does not know
the answer already.}
}
\value{
The number of features, or \code{NA} when the driver would have to scan
the layer and \code{force} is \code{FALSE}. Any filter set on the layer applies.
}
\description{
Count the features of a layer
}
\examples{
ds <- gdal_open(system.file("extdata/test.gpkg", package = "GDAL7"))
feature_count(get_layer(ds, 1))
gdal_close(ds)
}
