% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{flush_cache}
\alias{flush_cache}
\title{FlushCache}
\usage{
flush_cache(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\value{
integer
}
\description{
FlushCache
}
