% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAL7-package.R
\docType{package}
\name{GDAL7-package}
\alias{GDAL7}
\alias{GDAL7-package}
\title{GDAL7: Modern GDAL Bindings for R Using S7}
\description{
Comprehensive GDAL bindings for R using S7 classes and cpp11. Auto-generated from GDAL's SWIG interface definitions.
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
