% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{gdal_get_driver}
\alias{gdal_get_driver}
\title{Get a driver by index}
\usage{
gdal_get_driver(index)
}
\arguments{
\item{index}{Driver index (0-based)}
}
\value{
A GDALDriver object, or NULL if invalid index
}
\description{
Get a driver by index
}
