% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{gdal_get_driver_count}
\alias{gdal_get_driver_count}
\title{Get the number of registered GDAL drivers}
\usage{
gdal_get_driver_count()
}
\value{
Integer number of drivers
}
\description{
Get the number of registered GDAL drivers
}
