% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{GDALDataset}
\alias{GDALDataset}
\title{GDAL Dataset class}
\usage{
GDALDataset(.ptr = NULL)
}
\arguments{
\item{.ptr}{Internal. External pointer to the underlying GDAL object.}
}
\description{
S7 class wrapping GDALDataset
}
