% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{GDALDriver}
\alias{GDALDriver}
\title{GDALDriver class}
\usage{
GDALDriver(.ptr = NULL)
}
\arguments{
\item{.ptr}{Internal. External pointer to the underlying GDAL object.}
}
\description{
Represents a GDAL format driver (e.g., GTiff, GPKG, etc.)
}
