% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-multidim.R
\name{GDALGroup}
\alias{GDALGroup}
\title{GDALGroup class}
\usage{
GDALGroup(.ptr = NULL)
}
\arguments{
\item{.ptr}{Internal. External pointer to the underlying GDAL object.}
}
\description{
Represents a group in a multidimensional dataset (NetCDF, Zarr, HDF5, etc.)
Groups can contain arrays (GDALMDArray) and subgroups.
}
