% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-majorobject.R
\name{GDALMajorObject}
\alias{GDALMajorObject}
\title{GDAL MajorObject class}
\usage{
GDALMajorObject(.ptr = NULL)
}
\arguments{
\item{.ptr}{Internal. External pointer to the underlying GDAL object.}
}
\description{
S7 class wrapping GDALMajorObject
}
