% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{GDALRasterBand}
\alias{GDALRasterBand}
\title{GDALRasterBand class}
\usage{
GDALRasterBand(.ptr = NULL)
}
\arguments{
\item{.ptr}{Internal. External pointer to the underlying GDAL object.}
}
\description{
Represents a raster band within a GDAL dataset.
}
