% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{get_band_number}
\alias{get_band_number}
\title{Get band number}
\usage{
get_band_number(x, ...)
}
\arguments{
\item{x}{A GDALRasterBand object}

\item{...}{Arguments passed on to methods.}
}
\value{
Integer band number (1-indexed)
}
\description{
Get band number
}
