% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{get_block_size}
\alias{get_block_size}
\title{Get band block size (tile dimensions)}
\usage{
get_block_size(x, ...)
}
\arguments{
\item{x}{A GDALRasterBand object}

\item{...}{Arguments passed on to methods.}
}
\value{
Named integer vector with 'x' and 'y' block dimensions
}
\description{
Get band block size (tile dimensions)
}
