% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{get_close_reports_progress}
\alias{get_close_reports_progress}
\title{GetCloseReportsProgress}
\usage{
get_close_reports_progress(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\value{
logical
}
\description{
GetCloseReportsProgress
}
