% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{get_color_interpretation}
\alias{get_color_interpretation}
\title{Get band color interpretation (numeric code)}
\usage{
get_color_interpretation(x, ...)
}
\arguments{
\item{x}{A GDALRasterBand object}

\item{...}{Arguments passed on to methods.}
}
\value{
Integer color interpretation code
}
\description{
Get band color interpretation (numeric code)
}
