% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{get_color_interpretation_name}
\alias{get_color_interpretation_name}
\title{Get band color interpretation name}
\usage{
get_color_interpretation_name(x, ...)
}
\arguments{
\item{x}{A GDALRasterBand object}

\item{...}{Arguments passed on to methods.}
}
\value{
Character color interpretation name (e.g., "Gray", "Red", "Green", "Blue")
}
\description{
Get band color interpretation name
}
