% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-multidim.R
\name{get_coordinate_variables}
\alias{get_coordinate_variables}
\title{Get the coordinate variables of an array}
\usage{
get_coordinate_variables(x, ...)
}
\arguments{
\item{x}{A GDALMDArray object}

\item{...}{Arguments passed on to methods.}
}
\value{
A list of GDALMDArray objects, empty when the format names none.
}
\description{
The arrays a format names as this one's coordinates. This is not quite the
same question as which dimensions are indexed: a swath carries latitude and
longitude arrays that are two-dimensional and index no dimension at all. For
the ordinary gridded case, \code{\link[=get_dimension_values]{get_dimension_values()}} is the shorter road.
}
