% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{get_creation_options}
\alias{get_creation_options}
\title{Get driver creation options XML}
\usage{
get_creation_options(x, ...)
}
\arguments{
\item{x}{A GDALDriver object}

\item{...}{Arguments passed on to methods.}
}
\value{
Character XML string describing creation options
}
\description{
Returns the XML describing creation options for this driver.
}
