% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{get_data_type}
\alias{get_data_type}
\title{Get band data type (numeric code)}
\usage{
get_data_type(x, ...)
}
\arguments{
\item{x}{A GDALRasterBand object}

\item{...}{Arguments passed on to methods.}
}
\value{
Integer data type code (see GDAL GDALDataType enum)
}
\description{
Get band data type (numeric code)
}
