% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{get_data_type_name}
\alias{get_data_type_name}
\title{Get band data type name}
\usage{
get_data_type_name(x, ...)
}
\arguments{
\item{x}{A GDALRasterBand object}

\item{...}{Arguments passed on to methods.}
}
\value{
Character data type name (e.g., "Byte", "Int16", "Float32")
}
\description{
Get band data type name
}
