% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-majorobject.R
\name{get_description}
\alias{get_description}
\title{GetDescription}
\usage{
get_description(x, ...)
}
\arguments{
\item{x}{A GDALMajorObject object}

\item{...}{Arguments passed on to methods.}
}
\value{
character
}
\description{
GetDescription
}
