% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-multidim.R
\name{get_dimension_count}
\alias{get_dimension_count}
\title{Get number of dimensions in an array}
\usage{
get_dimension_count(x, ...)
}
\arguments{
\item{x}{A GDALMDArray object}

\item{...}{Arguments passed on to methods.}
}
\value{
Integer dimension count
}
\description{
Get number of dimensions in an array
}
