% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{get_driver}
\alias{get_driver}
\title{Get the driver that opened a dataset}
\usage{
get_driver(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\value{
A GDALDriver object, or NULL
}
\description{
Get the driver that opened a dataset
}
