% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{get_field_domain_names}
\alias{get_field_domain_names}
\title{GetFieldDomainNames}
\usage{
get_field_domain_names(x, options = NULL)
}
\arguments{
\item{x}{A GDALDataset object}

\item{options}{character (default: NULL)}
}
\value{
character
}
\description{
GetFieldDomainNames
}
