% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{get_file_list}
\alias{get_file_list}
\title{GetFileList}
\usage{
get_file_list(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\value{
character
}
\description{
GetFileList
}
