% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-multidim.R
\name{get_full_name}
\alias{get_full_name}
\title{Get group full name (path)}
\usage{
get_full_name(x, ...)
}
\arguments{
\item{x}{A GDALGroup object}

\item{...}{Arguments passed on to methods.}
}
\value{
Character full path name
}
\description{
Get group full name (path)
}
