% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{get_gcpcount}
\alias{get_gcpcount}
\title{GetGCPCount}
\usage{
get_gcpcount(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\value{
integer
}
\description{
GetGCPCount
}
