% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{get_gcpprojection}
\alias{get_gcpprojection}
\title{GetGCPProjection}
\usage{
get_gcpprojection(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\value{
character
}
\description{
GetGCPProjection
}
