% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-multidim.R
\name{get_group_names}
\alias{get_group_names}
\title{Get names of subgroups in a group}
\usage{
get_group_names(x, ...)
}
\arguments{
\item{x}{A GDALGroup object}

\item{...}{Arguments passed on to methods.}
}
\value{
Character vector of subgroup names
}
\description{
Get names of subgroups in a group
}
