% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{get_help_topic}
\alias{get_help_topic}
\title{Get driver help topic URL}
\usage{
get_help_topic(x, ...)
}
\arguments{
\item{x}{A GDALDriver object}

\item{...}{Arguments passed on to methods.}
}
\value{
Character help topic URL
}
\description{
Get driver help topic URL
}
