% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{get_layer}
\alias{get_layer}
\title{Get a layer from a dataset}
\usage{
get_layer(x, layer = 1)
}
\arguments{
\item{x}{A GDALDataset object.}

\item{layer}{The layer to take: its name, or its position counting from 1.}
}
\value{
A GDALLayer object.
}
\description{
Get a layer from a dataset
}
\examples{
ds <- gdal_open(system.file("extdata/test.gpkg", package = "GDAL7"))
get_layer(ds, 1)
gdal_close(ds)
}
