% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{get_layer_count}
\alias{get_layer_count}
\title{GetLayerCount}
\usage{
get_layer_count(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\value{
integer
}
\description{
GetLayerCount
}
