% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{get_long_name}
\alias{get_long_name}
\title{Get driver long name}
\usage{
get_long_name(x, ...)
}
\arguments{
\item{x}{A GDALDriver object}

\item{...}{Arguments passed on to methods.}
}
\value{
Character driver long name (e.g., "GeoTIFF", "GeoPackage")
}
\description{
Get driver long name
}
