% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-multidim.R
\name{get_mdarray_names}
\alias{get_mdarray_names}
\title{Get names of arrays in a group}
\usage{
get_mdarray_names(x, ...)
}
\arguments{
\item{x}{A GDALGroup object}

\item{...}{Arguments passed on to methods.}
}
\value{
Character vector of array names
}
\description{
Get names of arrays in a group
}
