% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-majorobject.R
\name{get_metadata_dict}
\alias{get_metadata_dict}
\title{GetMetadata_Dict}
\usage{
get_metadata_dict(x, pszDomain = "")
}
\arguments{
\item{x}{A GDALMajorObject object}

\item{pszDomain}{character (default: )}
}
\value{
character
}
\description{
GetMetadata_Dict
}
