% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-majorobject.R
\name{get_metadata_domain_list}
\alias{get_metadata_domain_list}
\title{GetMetadataDomainList}
\usage{
get_metadata_domain_list(x, ...)
}
\arguments{
\item{x}{A GDALMajorObject object}

\item{...}{Arguments passed on to methods.}
}
\value{
character
}
\description{
GetMetadataDomainList
}
