% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-majorobject.R
\name{get_metadata_item}
\alias{get_metadata_item}
\title{GetMetadataItem}
\usage{
get_metadata_item(x, pszName, pszDomain = "")
}
\arguments{
\item{x}{A GDALMajorObject object}

\item{pszName}{ANY}

\item{pszDomain}{ANY (default: )}
}
\value{
character
}
\description{
GetMetadataItem
}
