% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-majorobject.R
\name{get_metadata_list}
\alias{get_metadata_list}
\title{GetMetadata_List}
\usage{
get_metadata_list(x, pszDomain = "")
}
\arguments{
\item{x}{A GDALMajorObject object}

\item{pszDomain}{character (default: "")}
}
\value{
character
}
\description{
GetMetadata_List
}
