% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-multidim.R
\name{get_name}
\alias{get_name}
\title{Get group name}
\usage{
get_name(x, ...)
}
\arguments{
\item{x}{A GDALGroup object}

\item{...}{Arguments passed on to methods.}
}
\value{
Character group name
}
\description{
Get group name
}
