% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{get_nodata_value}
\alias{get_nodata_value}
\title{Get band nodata value}
\usage{
get_nodata_value(x, ...)
}
\arguments{
\item{x}{A GDALRasterBand object}

\item{...}{Arguments passed on to methods.}
}
\value{
Numeric nodata value, or NULL if not set
}
\description{
Get band nodata value
}
