% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{get_offset}
\alias{get_offset}
\title{Get band offset value}
\usage{
get_offset(x, ...)
}
\arguments{
\item{x}{A GDALRasterBand object}

\item{...}{Arguments passed on to methods.}
}
\value{
Numeric offset value, or NULL if not set
}
\description{
Get band offset value
}
