% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-io.R
\name{get_overview_count}
\alias{get_overview_count}
\title{Number of overviews on a band}
\usage{
get_overview_count(x, ...)
}
\arguments{
\item{x}{A GDALRasterBand object}

\item{...}{Arguments passed on to methods.}
}
\value{
Integer count.
}
\description{
Number of overviews on a band
}
