% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{get_projection}
\alias{get_projection}
\title{GetProjection}
\usage{
get_projection(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\value{
character
}
\description{
GetProjection
}
