% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{get_projection_ref}
\alias{get_projection_ref}
\title{GetProjectionRef}
\usage{
get_projection_ref(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\value{
character
}
\description{
GetProjectionRef
}
