% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{get_raster_band}
\alias{get_raster_band}
\title{Get a raster band from a dataset}
\usage{
get_raster_band(x, nBand)
}
\arguments{
\item{x}{A GDALDataset object}

\item{nBand}{Band number, 1-indexed}
}
\value{
A GDALRasterBand object
}
\description{
Get a raster band from a dataset
}
