% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{get_raster_count}
\alias{get_raster_count}
\title{Get number of raster bands}
\usage{
get_raster_count(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\value{
Integer number of bands
}
\description{
Get number of raster bands
}
