% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{get_raster_xsize}
\alias{get_raster_xsize}
\title{Get raster width}
\usage{
get_raster_xsize(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\value{
Integer width in pixels
}
\description{
Get raster width
}
