% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{get_raster_ysize}
\alias{get_raster_ysize}
\title{Get raster height}
\usage{
get_raster_ysize(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\value{
Integer height in pixels
}
\description{
Get raster height
}
