% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{get_relationship_names}
\alias{get_relationship_names}
\title{GetRelationshipNames}
\usage{
get_relationship_names(x, options = NULL)
}
\arguments{
\item{x}{A GDALDataset object}

\item{options}{character (default: NULL)}
}
\value{
character
}
\description{
GetRelationshipNames
}
