% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-multidim.R
\name{get_root_group}
\alias{get_root_group}
\title{Get root group from a multidimensional dataset}
\usage{
get_root_group(x, ...)
}
\arguments{
\item{x}{A GDALDataset opened with multidim = TRUE}

\item{...}{Arguments passed on to methods.}
}
\value{
A GDALGroup object, or NULL if not a multidimensional dataset
}
\description{
Get root group from a multidimensional dataset
}
