% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{get_scale}
\alias{get_scale}
\title{Get band scale factor}
\usage{
get_scale(x, ...)
}
\arguments{
\item{x}{A GDALRasterBand object}

\item{...}{Arguments passed on to methods.}
}
\value{
Numeric scale factor, or NULL if not set
}
\description{
Get band scale factor
}
