% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{get_short_name}
\alias{get_short_name}
\title{Get driver short name}
\usage{
get_short_name(x, ...)
}
\arguments{
\item{x}{A GDALDriver object}

\item{...}{Arguments passed on to methods.}
}
\value{
Character driver short name (e.g., "GTiff", "GPKG")
}
\description{
Get driver short name
}
