% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{get_spatial_ref}
\alias{get_spatial_ref}
\title{GetSpatialRef}
\usage{
get_spatial_ref(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\value{
OSRSpatialReference
}
\description{
GetSpatialRef
}
