% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{get_unit_type}
\alias{get_unit_type}
\title{Get band unit type}
\usage{
get_unit_type(x, ...)
}
\arguments{
\item{x}{A GDALRasterBand object}

\item{...}{Arguments passed on to methods.}
}
\value{
Character unit type string (e.g., "meters", "degrees")
}
\description{
Get band unit type
}
