% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{get_xsize}
\alias{get_xsize}
\title{Get band width}
\usage{
get_xsize(x, ...)
}
\arguments{
\item{x}{A GDALRasterBand object}

\item{...}{Arguments passed on to methods.}
}
\value{
Integer width in pixels
}
\description{
Get band width
}
