% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{get_ysize}
\alias{get_ysize}
\title{Get band height}
\usage{
get_ysize(x, ...)
}
\arguments{
\item{x}{A GDALRasterBand object}

\item{...}{Arguments passed on to methods.}
}
\value{
Integer height in pixels
}
\description{
Get band height
}
