% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{has_open_option}
\alias{has_open_option}
\title{Check if driver has an open option}
\usage{
has_open_option(x, option_name)
}
\arguments{
\item{x}{A GDALDriver object}

\item{option_name}{Name of the open option to check}
}
\value{
Logical TRUE if driver supports the option
}
\description{
Check if driver has an open option
}
