% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{is_thread_safe}
\alias{is_thread_safe}
\title{IsThreadSafe}
\usage{
is_thread_safe(x, nScopeFlags)
}
\arguments{
\item{x}{A GDALDataset object}

\item{nScopeFlags}{integer}
}
\value{
logical
}
\description{
IsThreadSafe
}
