% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{mark_suppress_on_close}
\alias{mark_suppress_on_close}
\title{MarkSuppressOnClose}
\usage{
mark_suppress_on_close(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\description{
MarkSuppressOnClose
}
