% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-multidim.R
\name{open_group}
\alias{open_group}
\title{Open a subgroup from a group}
\usage{
open_group(x, name)
}
\arguments{
\item{x}{A GDALGroup object}

\item{name}{Subgroup name}
}
\value{
A GDALGroup object, or NULL if not found
}
\description{
Open a subgroup from a group
}
