% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{reset_reading}
\alias{reset_reading}
\title{ResetReading}
\usage{
reset_reading(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\description{
ResetReading
}
