% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{rollback_transaction}
\alias{rollback_transaction}
\title{RollbackTransaction}
\usage{
rollback_transaction(x, ...)
}
\arguments{
\item{x}{A GDALDataset object}

\item{...}{Arguments passed on to methods.}
}
\value{
integer
}
\description{
RollbackTransaction
}
