% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-majorobject.R
\name{set_description}
\alias{set_description}
\title{SetDescription}
\usage{
set_description(x, pszNewDesc)
}
\arguments{
\item{x}{A GDALMajorObject object}

\item{pszNewDesc}{ANY}
}
\description{
SetDescription
}
