% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-io.R
\name{set_geotransform}
\alias{set_geotransform}
\title{Set the geotransform of a dataset}
\usage{
set_geotransform(x, geotransform)
}
\arguments{
\item{x}{A GDALDataset object, opened for update.}

\item{geotransform}{A numeric vector of length 6, in GDAL's order.}
}
\value{
\code{x}, invisibly.
}
\description{
Set the geotransform of a dataset
}
