% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-majorobject.R
\name{set_metadata_2}
\alias{set_metadata_2}
\title{SetMetadata (overload 2)}
\usage{
set_metadata_2(x, pszMetadataString, pszDomain = "")
}
\arguments{
\item{x}{A GDALMajorObject object}

\item{pszMetadataString}{character}

\item{pszDomain}{character}
}
\value{
integer
}
\description{
SetMetadata (overload 2)
}
