% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{set_projection}
\alias{set_projection}
\title{SetProjection}
\usage{
set_projection(x, prj)
}
\arguments{
\item{x}{A GDALDataset object}

\item{prj}{character}
}
\value{
integer
}
\description{
SetProjection
}
