% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{start_transaction}
\alias{start_transaction}
\title{StartTransaction}
\usage{
start_transaction(x, force = FALSE)
}
\arguments{
\item{x}{A GDALDataset object}

\item{force}{integer (default: FALSE)}
}
\value{
integer
}
\description{
StartTransaction
}
