% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-multidim.R
\name{GDALMDArray}
\alias{GDALMDArray}
\title{GDALMDArray class}
\usage{
GDALMDArray(.ptr = NULL)
}
\arguments{
\item{.ptr}{Internal. External pointer to the underlying GDAL object.}
}
\description{
Represents a multidimensional array in a GDAL dataset.
}
