% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-multidim.R
\name{get_dimensions}
\alias{get_dimensions}
\title{Get dimensions of an array}
\usage{
get_dimensions(x, ...)
}
\arguments{
\item{x}{A GDALMDArray object}

\item{...}{Arguments passed on to methods.}
}
\value{
Data frame with dimension names and sizes
}
\description{
Get dimensions of an array
}
