% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-io.R
\name{get_overview}
\alias{get_overview}
\title{Get one overview of a band}
\usage{
get_overview(x, index)
}
\arguments{
\item{x}{A GDALRasterBand object}

\item{index}{Zero-based overview index, as GDAL numbers them.}
}
\value{
A GDALRasterBand object for the overview, which belongs to the same
dataset as \code{x}.
}
\description{
Get one overview of a band
}
