% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-info.R
\name{get_raster_band}
\alias{get_raster_band}
\title{Get a raster band from a dataset}
\usage{
get_raster_band(x, nBand)
}
\arguments{
\item{x}{A GDALDataset object}

\item{nBand}{Band number, 1-indexed}
}
\value{
A GDALRasterBand object
}
\description{
A band is taken by number, so this is a function rather than a property.
What the band then knows about itself is a property of the band; see
\link{GDALRasterBand}.
}
\examples{
ds <- gdal_open(system.file("extdata/test.tif", package = "GDAL7"))
get_raster_band(ds, 1)
gdal_close(ds)
}
