% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-class-dataset.R
\name{get_thread_safe_dataset}
\alias{get_thread_safe_dataset}
\title{GetThreadSafeDataset}
\usage{
get_thread_safe_dataset(x, nScopeFlags)
}
\arguments{
\item{x}{A GDALDataset object}

\item{nScopeFlags}{integer}
}
\value{
GDALDataset
}
\description{
GetThreadSafeDataset
}
