% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-multidim.R
\name{open_mdarray}
\alias{open_mdarray}
\title{Open an array from a group}
\usage{
open_mdarray(x, name)
}
\arguments{
\item{x}{A GDALGroup object}

\item{name}{The array's name within this group, or a path from the root of
the dataset such as \code{"/weather/temperature"}.}
}
\value{
A GDALMDArray object, or NULL if not found
}
\description{
Open an array from a group
}
