% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-majorobject.R
\name{set_metadata}
\alias{set_metadata}
\title{SetMetadata}
\usage{
set_metadata(x, papszMetadata = NULL, pszDomain = "")
}
\arguments{
\item{x}{A GDALMajorObject object}

\item{papszMetadata}{character (default: NULL)}

\item{pszDomain}{character (default: "")}
}
\value{
integer
}
\description{
SetMetadata
}
