% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-majorobject.R
\name{set_metadata_2}
\alias{set_metadata_2}
\title{SetMetadata, single string overload}
\usage{
set_metadata_2(x, pszMetadataString, pszDomain = "")
}
\arguments{
\item{x}{A GDALMajorObject object}

\item{pszMetadataString}{character. A single "NAME=VALUE" string.}

\item{pszDomain}{character. Metadata domain, "" for the default domain.}
}
\value{
integer
}
\description{
SetMetadata, single string overload
}
