% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-class-majorobject.R
\name{set_metadata_item}
\alias{set_metadata_item}
\title{SetMetadataItem}
\usage{
set_metadata_item(x, pszName, pszValue, pszDomain = "")
}
\arguments{
\item{x}{A GDALMajorObject object}

\item{pszName}{character}

\item{pszValue}{character}

\item{pszDomain}{character (default: )}
}
\value{
integer
}
\description{
SetMetadataItem
}
